Overhaulled routes.js and added controllers
- Still need to clean up each controller and add an ESI controller.
